Finish iOS review notification test split
